# src/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Evently",
    "termsOfService": "",
    "contact": {
      "email": "evently@localhost:5000"
    },
    "license": {
      "name": "evently",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "https://evently-backend.herokuapp.com/",
  "basePath": "/",
...
  "schemes": ["https", "http"],
  "paths": {
...
    "api/ticket/past": {
      "get": {
        "tags": ["ticket"],
        "summary": "fetch all user's past tickets",
        "description": "",
        "operationId": "past tickets",
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Add the token",
            "type": "string",
            "required": true,
            "example": "Bearer token"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Eventstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://Eventstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:Events": "modify Events in your account",
        "read:Events": "read your Events"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
...
}